style: NewLineService의 이름을 LineService로 변경
