--- src/main/java/wooteco/subway/admin/dto/LineStationCreateRequest.java
@@ -1,34 +1,0 @@
-package wooteco.subway.admin.dto;
-
-public class LineStationCreateRequest {
-	private Long preStationId;
-	private Long stationId;
-	private int distance;
-	private int duration;
-
-	public LineStationCreateRequest() {
-	}
-
-	public LineStationCreateRequest(Long preStationId, Long stationId, int distance, int duration) {
-		this.preStationId = preStationId;
-		this.stationId = stationId;
-		this.distance = distance;
-		this.duration = duration;
-	}
-
-	public Long getPreStationId() {
-		return preStationId;
-	}
-
-	public Long getStationId() {
-		return stationId;
-	}
-
-	public int getDistance() {
-		return distance;
-	}
-
-	public int getDuration() {
-		return duration;
-	}
-}
