+++ src/test/java/wooteco/subway/admin/service/LineServiceTest.java
@@ -1,14 +1,11 @@
 package wooteco.subway.admin.service;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
 import java.time.LocalTime;
 import java.util.Optional;
-import java.util.Set;
 
-import org.assertj.core.util.Sets;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -17,9 +14,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import wooteco.subway.admin.domain.Line;
 import wooteco.subway.admin.domain.LineStation;
-import wooteco.subway.admin.domain.Station;
-import wooteco.subway.admin.dto.LineResponse;
-import wooteco.subway.admin.dto.LineStationCreateRequest;
 import wooteco.subway.admin.repository.LineRepository;
 import wooteco.subway.admin.repository.StationRepository;
 
@@ -113,17 +107,5 @@ public class LineServiceTest {
 		assertThat(line.getStations()).hasSize(2);
 		assertThat(line.getLineStationsId().get(0)).isEqualTo(1L);
 		assertThat(line.getLineStationsId().get(1)).isEqualTo(2L);
-	}
-
-	@Test
-	void findLineWithStationsById() {
-		Set<Station> stations = Sets.newLinkedHashSet(new Station("강남역"), new Station("역삼역"),
-				new Station("삼성역"));
-		when(lineRepository.findById(anyLong())).thenReturn(Optional.of(line));
-		when(stationRepository.findAllById(anyList())).thenReturn(stations);
-
-		LineResponse lineResponse = lineService.findLineWithStationsById(1L);
-
-		assertThat(lineResponse.getStations()).hasSize(3);
 	}
 }
